+++ app/src/main/java/woowacourse/movie/feature/model/seat/SeatUiModel.kt
@@ -1,0 +1,10 @@
+package woowacourse.movie.feature.model.seat
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class SeatUiModel(
+    val row: Int,
+    val col: Int,
+) : Parcelable
