+++ app/src/main/java/woowacourse/movie/feature/model/seat/SeatsUiModel.kt
@@ -1,0 +1,10 @@
+package woowacourse.movie.feature.model.seat
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+class SeatsUiModel(
+    val capacity: Int,
+    val seats: List<SeatUiModel>,
+) : Parcelable
