+++ app/src/main/java/woowacourse/movie/feature/movieSelection/MovieSelectionContract.kt
@@ -1,0 +1,18 @@
+package woowacourse.movie.feature.movieSelection
+
+import woowacourse.movie.feature.model.movie.MovieListItem
+import woowacourse.movie.feature.model.movie.MovieListItem.MovieUiModel
+
+interface MovieSelectionContract {
+    interface View {
+        fun showMovies(movies: List<MovieListItem>)
+
+        fun goToReservation(movie: MovieUiModel)
+    }
+
+    interface Presenter {
+        fun onViewCreated()
+
+        fun onMovieSelection(movie: MovieUiModel)
+    }
+}
