--- app/src/main/java/woowacourse/movie/view/model/movie/TicketUiModel.kt
@@ -1,13 +1,0 @@
-package woowacourse.movie.view.model.movie
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-import woowacourse.movie.view.model.movie.MovieListItem.MovieUiModel
-import java.time.LocalDateTime
-
-@Parcelize
-data class TicketUiModel(
-    val movie: MovieUiModel,
-    val showtime: LocalDateTime,
-    val count: Int,
-) : Parcelable
