--- app/src/main/java/woowacourse/movie/view/model/seat/SeatUiModel.kt
@@ -1,10 +1,0 @@
-package woowacourse.movie.view.model.seat
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class SeatUiModel(
-    val row: Int,
-    val col: Int,
-) : Parcelable
