--- app/src/main/java/woowacourse/movie/view/model/seat/SeatsUiModel.kt
@@ -1,10 +1,0 @@
-package woowacourse.movie.view.model.seat
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-class SeatsUiModel(
-    val capacity: Int,
-    val seats: List<SeatUiModel>,
-) : Parcelable
