+++ app/src/main/res/layout/activity_movie_reservation_result.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.movieReservationResult.MovieReservationResultActivity">
+    tools:context=".feature.movieReservationResult.MovieReservationResultActivity">
 
     <TextView
         android:id="@+id/cancellation_guide"
