+++ app/src/main/res/layout/activity_movie_selection.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.movieSelection.MovieSelectionActivity">
+    tools:context=".feature.movieSelection.MovieSelectionActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/movie_list"
