+++ app/src/test/java/woowacourse/movie/feature/MovieReservationResultPresenterTest.kt
@@ -1,0 +1,35 @@
+package woowacourse.movie.feature
+
+import io.mockk.mockk
+import io.mockk.verifyAll
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import woowacourse.movie.feature.movieReservationResult.MovieReservationResultContract
+import woowacourse.movie.feature.movieReservationResult.MovieReservationResultPresenter
+import woowacourse.movie.fixtures.seats
+import woowacourse.movie.fixtures.ticket
+
+class MovieReservationResultPresenterTest {
+    private lateinit var presenter: MovieReservationResultPresenter
+    private lateinit var view: MovieReservationResultContract.View
+
+    @BeforeEach
+    fun setup() {
+        view = mockk(relaxed = true)
+        presenter = MovieReservationResultPresenter(view)
+    }
+
+    @Test
+    fun `실행 시 예매 완료 정보가 표시된다`() {
+        // when
+        presenter.onViewCreated(ticket, seats)
+
+        // then
+        verifyAll {
+            view.showMovieDateTime(any())
+            view.showTicketCount(any())
+            view.showSelectedSeats(any())
+            view.showTotalPrice(any())
+        }
+    }
+}
