--- app/src/test/java/woowacourse/movie/presenter/MovieSelectionPresenterTest.kt
@@ -1,44 +1,0 @@
-package woowacourse.movie.presenter
-
-import io.mockk.Runs
-import io.mockk.every
-import io.mockk.just
-import io.mockk.mockk
-import io.mockk.verify
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-import woowacourse.movie.fixtures.movie
-import woowacourse.movie.presenter.movieSelection.MovieSelectionContract
-import woowacourse.movie.presenter.movieSelection.MovieSelectionPresenter
-
-class MovieSelectionPresenterTest {
-    private lateinit var presenter: MovieSelectionPresenter
-    private lateinit var view: MovieSelectionContract.View
-
-    @BeforeEach
-    fun setup() {
-        view = mockk(relaxed = true)
-        presenter = MovieSelectionPresenter(view)
-    }
-
-    @Test
-    fun `실행 시 영화 목록을 표시한다`() {
-        // when
-        presenter.onViewCreated()
-
-        // then
-        verify { view.showMovies(any()) }
-    }
-
-    @Test
-    fun `영화 선택 시 영화 예매 화면으로 이동한다`() {
-        // given
-        every { view.goToReservation(any()) } just Runs
-
-        // when
-        presenter.onMovieSelection(movie)
-
-        // then
-        verify { view.goToReservation(any()) }
-    }
-}
